refactor(acp-bridge): lift BridgeOptions + introduce DaemonStatusProvider seam (#4175 PR 22b/2)
Design slice of #4175 Wave 5 PR 22b. Lifts the bridge's construction
contract (`BridgeOptions`) to `@qwen-code/acp-bridge` and introduces a
new `DaemonStatusProvider` interface so the bridge factory no longer
hard-imports daemon-host-specific helpers. The mechanical bulk lift of
`BridgeClient` + spawn factory + `createHttpAcpBridge` factory closure
(~3000 LOC) lands in PR 22b/3 against this stable contract.

What's lifted

- `BridgeOptions` interface (~150 LOC) from `httpAcpBridge.ts:189-325`
  → `acp-bridge/src/bridgeOptions.ts`. Adds new optional
  `statusProvider?: DaemonStatusProvider` field; everything else
  preserved verbatim.
- `buildDaemonPreflightCells` + `safeCheck` (~210 LOC) deleted from
  `httpAcpBridge.ts:4002-4214` → moved to new
  `cli/src/serve/daemonStatusProvider.ts`. Logic byte-for-byte
  identical (slot-by-slot allocation, `Promise.allSettled` fan-out,
  classified error cells); only the call-site changes from
  `await buildDaemonPreflightCells(boundWorkspace)` to
  `await opts.statusProvider?.getDaemonPreflightCells(boundWorkspace) ?? []`.

What's new

- `DaemonStatusProvider` interface in `acp-bridge/src/bridgeOptions.ts`:
  two methods (`getEnvStatus`, `getDaemonPreflightCells`) covering the
  full set of daemon-host cells the bridge currently delegates. Direct
  positional args (boundWorkspace + acpChannelLive); both methods
  return `Promise<>` so future async impls (config-file reads, network
  probes) get the seam without breaking the bridge. No abort signal —
  YAGNI; future async needs can pull in a typed extension. Single
  combined interface (vs split into two) because env + preflight are
  conceptually one daemon-host status surface and the production impl
  ships them as one factory.
- `createDaemonStatusProvider()` in `cli/src/serve/daemonStatusProvider.ts`:
  thin wrapper that calls `buildEnvStatusFromProcess` (still in
  `envSnapshot.ts`) and the moved `buildDaemonPreflightCells` body.
  Stateless; production callers can re-allocate per-request without
  performance cost.

Bridge factory body changes

- `getWorkspaceEnvStatus`: routes through `opts.statusProvider?.getEnvStatus(...)`.
  When `statusProvider` is omitted (Mode A in-process consumers, tests
  that don't care about env cells) the bridge returns an idle envelope
  `{ v, workspaceCwd, initialized: true, acpChannelLive, cells: [] }`
  matching the "queryable but empty" pattern PR 12 / 13 established for
  diagnostic routes.
- `getWorkspacePreflightStatus`: routes the daemon-half through
  `opts.statusProvider?.getDaemonPreflightCells(...) ?? []`. ACP-side
  cells (auth / mcp_discovery / skills / providers / tool_registry /
  egress) keep being fetched separately via the existing
  `requestWorkspaceStatus` extMethod path; the bridge stitches the
  two halves together as before.

Wiring

- `runQwenServe.ts` passes `statusProvider: createDaemonStatusProvider()`
  into the bridge factory by default.
- `httpAcpBridge.test.ts`: `makeBridge` helper now defaults to providing
  the production status provider so existing 4 env / preflight tests
  (which exercise the bridge's delegation path) keep seeing populated
  cells. Tests that want to exercise the no-provider idle fallback can
  override with `{ statusProvider: undefined }`.
- `acp-bridge/package.json`: adds `./bridgeOptions` subpath export.
- `acp-bridge/src/index.ts`: barrel re-exports `bridgeOptions.js`.

Backward compatibility

- `httpAcpBridge.ts` still re-exports `BridgeOptions` and
  `DaemonStatusProvider` types for callers via `from './httpAcpBridge.js'`
  (existing pattern from PR 22b/1). server.ts:31, runQwenServe.ts:16,
  workspaceAgents.ts, workspaceMemory.ts all keep resolving without
  churn.
- Zero `/capabilities`, route, SDK, or behavior changes when the
  default `runQwenServe` is in play (production wiring is identical
  to pre-PR; the default `createDaemonStatusProvider()` reproduces the
  pre-injection cell shapes byte-for-byte).
- Direct embeds / tests that omit `statusProvider` see new idle
  fallback semantics; this is the documented Mode A integration
  pattern and matches PR 12 / 13's "idle status is queryable" contract.

Verification

- `cd packages/core && npx tsc --build` clean
- `cd packages/acp-bridge && npx tsc --noEmit` clean
- `cd packages/cli && npx tsc --noEmit` clean (no errors in
  `src/serve/`; pre-existing tsconfig-excluded `mcp.test.ts` /
  `check-i18n.ts` errors unrelated to this PR)
- `cd packages/cli && npx vitest run src/serve/` — 701/701 pass
  (16 test files; 4 env / preflight tests reuse the production
  provider via updated `makeBridge` helper)
- `npm ci --dry-run` clean (no lockfile drift; new file imports
  resolve via existing `@qwen-code/qwen-code-core` and `@qwen-code/acp-bridge`
  workspace deps already present from PR 22b/1)

Remaining PR 22b checkpoints

- 22b/3 (mechanical bulk lift): `BridgeClient` (~400 LOC) +
  `defaultSpawnChannelFactory` (~250 LOC) + `createHttpAcpBridge`
  factory closure (~2240 LOC) + 5064-LOC test file move + final
  `httpAcpBridge.ts` shim. Zero new design decisions — factory
  consumes `DaemonStatusProvider` per the contract this PR froze.
- 22b' (small follow-up): `WorkspaceFileSystem` injection into
  `BridgeClient.writeTextFile/readTextFile`. Closes PR 18 ws.ts:613
  TOCTOU thread.
- #4299: `mapDomainErrorToErrorKind` regex → typed errors (parallel
  follow-up tracked from PR 22b/1 review).
