+++ packages/acp-bridge/src/index.ts
@@ -12,3 +12,4 @@ export * from './workspacePaths.js';
 export * from './status.js';
 export * from './bridgeErrors.js';
 export * from './bridgeTypes.js';
+export * from './bridgeOptions.js';
